+++ api/server/router/cluster/cluster_routes.go
@@ -17,16 +17,9 @@ limitations under the License.
 package cluster
 
 import (
-	"context"
-	"io"
-	"time"
-
 	"github.com/gin-gonic/gin"
-	"github.com/gorilla/websocket"
-	"k8s.io/klog/v2"
 
 	"github.com/caoyingjunz/pixiu/api/server/httputils"
-	"github.com/caoyingjunz/pixiu/pkg/client"
 	"github.com/caoyingjunz/pixiu/pkg/types"
 )
 
@@ -309,44 +302,6 @@ func (cr *clusterRouter) watchPodLog(c *gin.Context) {
 		httputils.SetFailed(c, r, err)
 		return
 	}
-
-	req := cr.c.Cluster().WatchPodLog(c, opts.Cluster, opts.Namespace, opts.Pod, logOpt.Container, logOpt.TailLines)
-	if req == nil {
-		klog.Errorf("failed to get request")
-		return
-	}
-
-	withTimeout, cancelFunc := context.WithTimeout(c, time.Minute*10)
-	defer cancelFunc()
-
-	reader, err := req.Stream(withTimeout)
-	if err != nil {
-		httputils.SetFailed(c, r, err)
-		return
-	}
-	defer reader.Close()
-
-	conn, err := client.WebsocketUpgrade.Upgrade(c.Writer, c.Request, nil)
-	if err != nil {
-		httputils.SetFailed(c, r, err)
-		return
-	}
-	defer conn.Close()
-
-	client.WebsocketStore.Add(opts.Cluster, "log", conn)
-	client.WebsocketUpgrade.Subprotocols = []string{c.Request.Header.Get("Sec-WebSocket-Protocol")}
-	wsClient := client.NewWsClient(conn, opts.Cluster, "log")
-	for {
-		buf := make([]byte, 1024)
-		n, err := reader.Read(buf)
-		if err != nil && err != io.EOF {
-			klog.Errorf("failed to read message: %v", err)
-			break
-		}
-		err = wsClient.Conn.WriteMessage(websocket.TextMessage, buf[0:n])
-		if err != nil {
-			klog.Errorf("failed to write message: %v", err)
-			break
-		}
-	}
+	// websocket
+	cr.c.Cluster().WatchPodLog(c, opts.Cluster, opts.Namespace, opts.Pod, logOpt.Container, logOpt.TailLines, c.Writer, c.Request)
 }
